// runtime/src/types/research-assets.ts
export type PatentRecord = {
  id: string
  year: string
  department: string
  teacherName: string
  appointmentType: string
  title: string
  patentType: string
  status: string
}

export type TransferRecord = {
  id: string
  year: string
  department: string
  teacherName: string
  appointmentType: string
  title: string
  transferType: string
  company: string
  amount: number
}